Relocate SafeDatacubeDumper into utils/serialise.py
