+++ datacube/scripts/product.py
@@ -19,7 +19,8 @@ import pandas as pd
 from datacube.index.index import Index
 from datacube.ui import click as ui
 from datacube.ui.click import cli
-from datacube.utils import read_documents, InvalidDocException, SafeDatacubeDumper
+from datacube.utils import read_documents, InvalidDocException
+from datacube.utils.serialise import SafeDatacubeDumper
 
 _LOG = logging.getLogger('datacube-product')
